Renamed tesseract folder
