+++ moveit_core/collision_detection_bullet/CMakeLists.txt
@@ -3,8 +3,8 @@ set(MOVEIT_LIB_NAME moveit_collision_detection_bullet)
 add_library(${MOVEIT_LIB_NAME}
   src/collision_robot_bullet.cpp
   src/collision_world_bullet.cpp
-  src/tesseract/bullet_utils.cpp
-  src/tesseract/bullet_discrete_bvh_manager.cpp
+  src/bullet_integration/bullet_utils.cpp
+  src/bullet_integration/bullet_discrete_bvh_manager.cpp
 )
 set_target_properties(${MOVEIT_LIB_NAME} PROPERTIES VERSION "${${PROJECT_NAME}_VERSION}")
