+++ moveit_core/collision_detection_bullet/include/moveit/collision_detection_bullet/bullet_integration/basic_types.h
@@ -18,8 +18,8 @@
 
 /* Author: Levi Armstrong */
 
-#ifndef MOVEIT_COLLISION_DETECTION_BULLET_TESSERACT_BASIC_TYPES_H_
-#define MOVEIT_COLLISION_DETECTION_BULLET_TESSERACT_BASIC_TYPES_H_
+#ifndef MOVEIT_COLLISION_DETECTION_BULLET_BULLET_INTEGRATION_BASIC_TYPES_H_
+#define MOVEIT_COLLISION_DETECTION_BULLET_BULLET_INTEGRATION_BASIC_TYPES_H_
 
 #include <Eigen/Core>
 #include <Eigen/Geometry>
@@ -103,4 +103,4 @@ struct ContactTestData
 
 }  // namespace collision_detection_bullet
 
-#endif  // MOVEIT_COLLISION_DETECTION_BULLET_TESSERACT_BASIC_TYPES_H_
+#endif  // MOVEIT_COLLISION_DETECTION_BULLET_BULLET_INTEGRATION_BASIC_TYPES_H_
