+++ moveit_core/collision_detection_bullet/src/collision_world_bullet.cpp
@@ -36,8 +36,8 @@
 
 #include <moveit/collision_detection_bullet/collision_world_bullet.h>
 #include <moveit/collision_detection_bullet/collision_detector_allocator_bullet.h>
-#include <moveit/collision_detection_bullet/tesseract/ros_tesseract_utils.h>
-#include <moveit/collision_detection_bullet/tesseract/contact_checker_common.h>
+#include <moveit/collision_detection_bullet/bullet_integration/ros_bullet_utils.h>
+#include <moveit/collision_detection_bullet/bullet_integration/contact_checker_common.h>
 #include <boost/bind.hpp>
 #include <bullet/btBulletCollisionCommon.h>
